+++ src/components/pages/stats/LifetimeStats.vue
@@ -4,6 +4,9 @@ import { ref, computed, onMounted } from 'vue';
 import { useUserStore } from 'src/stores/user.ts';
 const userStore = useUserStore();
 
+import { useWorkStore } from 'src/stores/work';
+const workStore = useWorkStore();
+
 import { TALLY_MEASURE_INFO, formatCountValue, formatCountCounter } from 'src/lib/tally.ts';
 
 import { getDayCounts, DayCount } from 'src/lib/api/stats.ts';
@@ -17,16 +20,32 @@ const dailyStats = computed(() => {
   return calculateDailyStats(dayCounts.value);
 });
 
+const workStartingBalances = computed(() => {
+  if(workStore.works === null) { return {}; }
+
+  const startingBalances = workStore.works.reduce((balances, work) => {
+    for(const measure of Object.keys(work.startingBalance)) {
+      if(!(measure in balances)) { balances[measure] = 0; }
+      balances[measure] += work.startingBalance[measure];
+    }
+
+    return balances;
+  }, {});
+
+  return startingBalances;
+});
+
 const measures = computed(() => {
   const measuresAvailable = Object.keys(dailyStats.value.days);
-  const startingBalanceMeasures = Object.keys(userStore.user.userSettings.lifetimeStartingBalance);
+  const lifetimeStartingBalanceMeasures = Object.keys(userStore.user.userSettings.lifetimeStartingBalance);
+  const workStartingBalanceMeasures = Object.keys(workStartingBalances.value);
   // this makes sure they're in a consistent order
-  return Object.keys(TALLY_MEASURE_INFO).filter(measure => measuresAvailable.includes(measure) || startingBalanceMeasures.includes(measure));
+  return Object.keys(TALLY_MEASURE_INFO).filter(measure => measuresAvailable.includes(measure) || lifetimeStartingBalanceMeasures.includes(measure) || workStartingBalanceMeasures.includes(measure));
 });
 
 const totals = computed(() => {
   return measures.value.reduce((obj, measure) => {
-    obj[measure] = (dailyStats.value.totals[measure] || 0) + (userStore.user.userSettings.lifetimeStartingBalance[measure] || 0);
+    obj[measure] = (dailyStats.value.totals[measure] || 0) + (userStore.user.userSettings.lifetimeStartingBalance[measure] || 0) + (workStartingBalances.value[measure] || 0);
     return obj;
   }, {});
 });
@@ -57,6 +76,7 @@ const breadcrumbs: MenuItem[] = [
 ];
 
 onMounted(() => {
+  workStore.populate();
   loadDayCounts();
 });
 </script>
@@ -66,7 +86,10 @@ onMounted(() => {
     :breadcrumbs="breadcrumbs"
   >
     <div class="max-w-screen-lg">
-      <SectionTitle title="Lifetime Totals" />
+      <SectionTitle
+        title="Grand Totals"
+        subtitle="All your progress in TrackBear, plus project starting balances and your lifetime starting balance"
+      />
       <div class="total-counts flex flex-wrap justify-evenly gap-2 mb-4">
         <StatTile
           v-for="measure in measures"