+++ CHANGELOG.md
@@ -12,7 +12,10 @@ Types of changes include:
 
 ## Upcoming/Unreleased
 
+- NEW: Leaderboards in fundraiser mode now show their totals.
 - FIXED: The streak counter for habit goals on the dashboard no longer dumps a bunch of unrelated text and instead now gives a cheery "X in a row!" counter like it's supposed to. (h/t CaitSidhe, sarah, Quinoafox)
+- FIXED: Lifetime stats now take into account starting balances from your projects as well as your lifetime starting balance in your Settings. (h/t RJ)
+- FIXED: Public profiles now also take into account starting balances from your projects; they accidentally did not before.
 
 ## 0.15.0
 - NEW: Public profiles have launched! You can now enable your public profile in Settings. Once you do, your public profile will be available at trackbear.app/@[your username]. Profiles list lifetime stats by default, and you can add goals and projects to it by going to the goal or project, clicking **Configure**, and toggling the switch under "Show on profile?".